Update JtsSpatialContext Usage and add Geom-cloning workaround
https://github.com/locationtech/spatial4j/issues/150

Ran into an issue when upgrading the dependencies where
dateline-wrapping was not getting handled correctly if you
convert a geometry to a shape multiple times.

There seems to be some bit of internal state in the JTS geometry
that's getting out of whack. Cloning seems to fix the issue,
but hopefully I can eventually get this sorted out and replaced
with something better.