Replace intersecting geohashes implementation with simpler version
I'd like to potentially revisit this in the future because it does
seem that the existing version is a bit faster, but it also seems to
be a little buggy and doesn't cover some edge cases like
origin/dateline and multi-polygons very well.

For now I'm going to put in another version which is closer to how
I've seen this algorithm in other contexts (like geopathy).